// src/app/models/user.model.ts
export class User {
    user_id: number;
    full_name: string;
    contact_no: number;
    email: string;
    addr_1: string;
    addr_2: string;
    city: string;
    state: string;
    pincode: number;
    land_area: number;
    land_addr: string;
    land_pincode: number;
    account_no: number;
    ifsc_code: string;
    aadhar: string;
    pan: string;
    trader_license: string;
    certificate: string;
    password: string;
    user_type: string;
    status: string;
    
}